Extract package commands to a separate class
